fix(labels): remove table row labels form
